test(research): extract equation quality tests
